adsr: Support "infinite" delay and sustain.
This is useful if you don't know how long a given note will be held for…
you can just trigger the note upon pressing the button and have it
progress to the "sustain" state, at which point it'll hold the note.

You can then call `adsr_continue` to "release" that note when you detect
the button release and have it decay.