adsr: Fix "infinite" delay/sustain.
Seems `~value` doesn't quite have the intended effect.  The intent was
that, being the 1s compliment of `value`, the expression would return
zero if `value` was all ones, and non-zero.

This doesn't quite work the way I'd like it to.  A comparison should be
the same number of CPU cycles, and is more readable, so we'll do that.