attiny861 port: Fix annoying pop on each note.
It seems when the NJM2113 is first powered up, it emits a pop noise.
Not much I can do about that.  The thought was I should power down the
amplifier when it is not in use to save power.

The outcome is on each note, it would power the amp up, and it would
pop.

Solution: create a 1msec timer for general timing events, and use that
to count down one minute before powering off the amplifier.  The toy
will pop on first power on, then if left alone for a minute, will pop
again on the first note, but will not pop so long as there's only short
gaps between notes.